token interface inference basic test case
