+++ detectors/token-interface-inference/Cargo.toml
@@ -11,6 +11,7 @@ clippy_utils = { workspace = true }
 dylint_linting = { workspace = true }
 if_chain = { workspace = true }
 utils = { workspace = true }
+edit-distance = "2.1.2"
 
 [package.metadata.rust-analyzer]
 rustc_private = true
