--- test-cases/token-interface-events/token-interface-events-2/vulnerable-example/Cargo.toml
@@ -1,39 +1,0 @@
-[package]
-name = "token-interface-events-vulnerable-2"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-soroban-sdk = "=21.4.0"
-soroban-token-sdk = { version = "21.4.0" }
-
-
-[dev_dependencies]
-soroban-sdk = { version = "=21.4.0", features = ["testutils"] }
-soroban-token-sdk = { version = "21.4.0" }
-
-
-[features]
-testutils = ["soroban-sdk/testutils"]
-
-[profile.release]
-opt-level = "z"
-overflow-checks = true
-debug = 0
-strip = "symbols"
-debug-assertions = false
-panic = "abort"
-codegen-units = 1
-lto = true
-
-[profile.release-with-logs]
-inherits = "release"
-debug-assertions = true
-
-[workspace.metadata.dylint]
-libraries = [
-    { path="/home/sofia/Documentos/scout/scout-soroban/detectors/token-interface-events" },
-]
