+++ test-cases/token-interface-inference/Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 exclude = [".cargo", "target"]
-members = ["token-interface-events-*/*"]
+members = ["token-interface-inference-*/*"]
 resolver = "2"
 
 [workspace.dependencies]
@@ -24,5 +24,5 @@ inherits = "release"
 
 [workspace.metadata.dylint]
 libraries = [
-    { path="/home/sofia/Documentos/scout/scout-soroban/detectors/token-interface-events" },
+    { path="/home/sofia/Documentos/scout/scout-soroban/detectors/token-interface-inference" },
 ]
