+++ test-cases/token-interface-inference/token-interface-inference-1/remediated-example/src/lib.rs
